+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - [#1460] Commandline import of text, images, shapes, symbol-class
 - [#1909] Export/import DefineBitsJPEG3/4s alpha channel to/from separate file
  ("PNG/GIF/JPEG+alpha" option in GUI, "-format image:png_gif_jpeg_alpha" for commandline)
+- [#1910] Copy/paste transform matrix to/from the clipboard
 
 ### Fixed
 - [#1904] NullPointerException when renaming invalid identifiers in AS1/2 files caused by missing charset
@@ -2778,6 +2779,7 @@ All notable changes to this project will be documented in this file.
 [#1690]: https://www.free-decompiler.com/flash/issues/1690
 [#1460]: https://www.free-decompiler.com/flash/issues/1460
 [#1909]: https://www.free-decompiler.com/flash/issues/1909
+[#1910]: https://www.free-decompiler.com/flash/issues/1910
 [#1904]: https://www.free-decompiler.com/flash/issues/1904
 [#595]: https://www.free-decompiler.com/flash/issues/595
 [#1908]: https://www.free-decompiler.com/flash/issues/1908